Implement quarkus-jte-extension
This is a very experimental version of the integration between jte and
quarkus.

I cannot lay down all the details of the implementation here about
how this was implemented as most of the work was done by Andreas,
but it is a usable version though probably not suitable for production yet.

Code brought from https://github.com/renannprado/quarkus-jte-example, but
with changes in the package names and tests.

Publishing to maven central was also setup.

Thanks @casid for helping to hack this extension.

Co-authored-by: Andreas Hager <andy@mazebert.com>